fix(map): fixes matching tiles and hill init

+++ Assets/src/scripts/entities/Tile.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 
 public class Tile
 {
-    public int X { get; set; }
-    public int Y { get; set; }
-    public int Z { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
+    public float Z { get; set; }
     public TileType Type { get; set; }
     public GameObject Obj { get; set; }
 
-    public Tile(int x, int y, int z)
+    public Tile(float x, float y, float z)
     {
         X = x;
         Y = y;
